# (perf) Refactor logger class to encapsulate bunyan instance; add requestLogger stress test