--- a/src/util/common/scrub.js
+++ b/src/util/common/scrub.js
@@ -1,5 +1,9 @@
 import hideSecrets from 'hide-secrets';
 
-export default function(obj, config = {}) {
-  return hideSecrets(obj, { badWords: config.scrubFields });
+export default function scrub(args, config = {}) {
+  if (Array.isArray(config.scrubFields) && config.scrubFields.length) {
+    return hideSecrets(args, { badWords: config.scrubFields });
+  }
+
+  return args;
 }