--- a/test/leakage/check.js
+++ b/test/leakage/check.js
@@ -3,7 +3,7 @@ import bunyan from 'bunyan';
 import Logger from '../..';
 
 const TIMEOUT = 20 * 1000;
-const ITERATIONS = 1000;
+const ITERATIONS = 600;
 
 describe('Heap Tests', function() {
   let log;
@@ -11,6 +11,27 @@ describe('Heap Tests', function() {
   describe('we-js-logger', () => {
     beforeEach(() => {
       log = new Logger({ stdout: false });
+    });
+
+    it(`does not leak when logging ${ITERATIONS} times`, function() {
+      this.timeout(TIMEOUT);
+      iterate(ITERATIONS, () => {
+        log.info({}, 'log time', {});
+      });
+    });
+
+    it(`does not leak when building and logging child ${ITERATIONS} times`, function() {
+      this.timeout(TIMEOUT);
+      iterate(ITERATIONS, () => {
+        const child = log.child();
+        child.info({}, 'child time', {});
+      });
+    });
+  });
+
+  describe('we-js-logger (with scrub fields)', () => {
+    beforeEach(() => {
+      log = new Logger({ stdout: false, scrubFields: ['foo'] });
     });
 
     it(`does not leak when logging ${ITERATIONS} times`, function() {