--- a/test/specs/util/scrub.spec.js
+++ b/test/specs/util/scrub.spec.js
@@ -1,0 +1,21 @@
+import scrub from '../../../src/util/common/scrub';
+
+// FIXME stub out `hideSecrets` -- doesn't seem to be working on first pass
+// import * as hideSecretsModule from 'hide-secrets';
+
+describe('util/common/scrub', () => {
+  let args;
+  beforeEach(() => {
+    args = ['test', 'args', 'are', 'so', { fun: 'yay' }];
+  });
+
+  it('returns args if no config.scrubFields are present', () => {
+    const result = scrub(args, { scrubFields: [] });
+    expect(result).to.eql(args);
+  });
+
+  it('passes args through hide-secrets if config.scrubFields are present', () => {
+    const result = scrub(args, { scrubFields: ['fun'] });
+    expect(result).to.eql(['test', 'args', 'are', 'so', { fun: '[SECRET]' }]);
+  });
+});